+++ rational/benches/primitive.rs
@@ -7,7 +7,7 @@ use criterion::{
 use dashu_base::Sign;
 use dashu_int::{IBig, UBig};
 use dashu_ratio::RBig;
-use rand_v08::prelude::*;
+use rand_v010::prelude::*;
 use std::ops::*;
 
 const SEED: u64 = 1;
@@ -16,10 +16,10 @@ fn random_rbig<R>(bits: usize, rng: &mut R) -> RBig
 where
     R: Rng + ?Sized,
 {
-    let sign = Sign::from(rng.gen_bool(0.5));
+    let sign = Sign::from(rng.random_bool(0.5));
     let numerator =
-        IBig::from_parts(sign, rng.gen_range(UBig::ONE << (bits - 1)..UBig::ONE << bits));
-    let denominator = rng.gen_range(UBig::ONE << (bits - 1)..UBig::ONE << bits);
+        IBig::from_parts(sign, rng.random_range(UBig::ONE << (bits - 1)..UBig::ONE << bits));
+    let denominator = rng.random_range(UBig::ONE << (bits - 1)..UBig::ONE << bits);
     RBig::from_parts(numerator, denominator)
 }
 
@@ -27,7 +27,7 @@ fn random_ubig<R>(bits: usize, rng: &mut R) -> UBig
 where
     R: Rng + ?Sized,
 {
-    rng.gen_range(UBig::ONE << (bits - 1)..UBig::ONE << bits)
+    rng.random_range(UBig::ONE << (bits - 1)..UBig::ONE << bits)
 }
 
 macro_rules! add_binop_benchmark {
@@ -70,7 +70,7 @@ fn rbig_reduction(criterion: &mut Criterion) {
         let common = random_ubig(bits, &mut rng);
         let n_mag = random_ubig(bits, &mut rng) * &common;
         let d = random_ubig(bits, &mut rng) * &common;
-        let n = IBig::from_parts(Sign::from(rng.gen_bool(0.5)), n_mag);
+        let n = IBig::from_parts(Sign::from(rng.random_bool(0.5)), n_mag);
         group.bench_with_input(
             BenchmarkId::from_parameter(format!("1e{}", log_bits)),
             &(n, d),